+++ src/operator/DivideService.java
@@ -12,7 +12,7 @@ public class DivideService implements OperatorService {
 	 
 	@Override
 	public void execute() {
-		int sum = dependency.execute()[0] / dependency.execute()[1]; 
+		double sum = dependency.execute()[0] / dependency.execute()[1]; 
 		
 		dependency.setValue(sum);
